Fix switch to full screen problem
