+++ app/src/main/AndroidManifest.xml
@@ -74,11 +74,6 @@
             android:screenOrientation="sensorLandscape" />
 
         <activity
-            android:name=".ui.activity.PlayActivity"
-            android:configChanges="screenSize|smallestScreenSize|screenLayout|orientation"
-            android:screenOrientation="sensorLandscape" />
-
-        <activity
             android:name=".ui.activity.SettingActivity"
             android:configChanges="screenSize|smallestScreenSize|screenLayout|orientation"
             android:screenOrientation="sensorLandscape" />
