--- app/src/main/res/layout/activity_play.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/black"
-    android:keepScreenOn="true">
-
-    <com.google.android.exoplayer2.ui.StyledPlayerView
-        android:id="@+id/video"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:animation_enabled="false"
-        app:auto_show="false"
-        app:controller_layout_id="@layout/view_controller_bottom"
-        app:resize_mode="fit" />
-
-    <include
-        android:id="@+id/progress"
-        layout="@layout/view_progress"
-        android:visibility="gone" />
-
-    <include
-        android:id="@+id/center"
-        layout="@layout/view_controller_center"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:visibility="gone" />
-
-</FrameLayout>
