+++ app/src/main/res/layout/view_controller_bottom.xml
@@ -51,7 +51,7 @@
                 android:textColor="@color/white" />
 
             <TextView
-                android:id="@+id/replay"
+                android:id="@+id/reset"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginEnd="16dp"
